Fix Activity Service Add accordian to home
Activity service fixed to really return an IActivity observable that
includes all the types passed in the array.

Created an accordian view in home that breaks things out by activities
Uses material accordian.

changed allActivities to not include angel which is a dormant activity type
(encompassed by investment with a divergent flag)